// StageV/src/app/services/models/invoice.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { TravelRequest } from '../models/travel-request';
export interface Invoice {
  hotelName?: string;
  hotelTotal?: number;
  id?: number;
  nights?: number;
  perDiemRate?: number;
  perDiemTotal?: number;
  pricePerNight?: number;
  ticketAmount?: number;
  ticketNumber?: string;
  totalAmount?: number;
  travelDays?: number;
  travelRequest?: TravelRequest;
  travelerEmail?: string;
  travelerFirstName?: string;
  travelerLastName?: string;
}