// StageV/src/app/services/models/travel-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Mission } from '../models/mission';
import { Project } from '../models/project';
import { User } from '../models/user';
export interface TravelRequest {
  departureDate?: string;
  destination?: string;
  duration?: number;
  id?: number;
  mission?: Mission;
  objective?: string;
  project?: Project;
  requester?: User;
  returnDate?: string;
  secondaryDestination?: string;
  status?: string;
  type?: string;
  version?: number;
  visaRequired?: boolean;
}